feat: introduce GabaritoCard component for rendering answer sheets
- Added a new `GabaritoCard` component to encapsulate the layout and styling of answer sheets.
- Refactored `GabaritoPreview` and `GabaritosBatchPreview` components to utilize the new `GabaritoCard`, improving code reusability and maintainability.
- Removed redundant styled components from `GabaritoPreview` and `GabaritosBatchPreview`, streamlining the codebase.
